Refactor shift_generation
- Update folder to more appropriate name
- Move demand-related code into separated file
- Move employee-loading to xml_loader
- Refactor code in shift_generation, especially by splitting up functions and reusing data
