+++ preprocessing/shift_generation.py
@@ -1,0 +1,315 @@
+from preprocessing.demand_processing import get_time_steps, get_time_periods, get_demand, \
+    combine_demand_intervals, combine_no_demand_intervals, get_start_events
+from utils.const import DESIRED_SHIFT_DURATION, ALLOWED_SHIFT_DURATION, TIME_DEFINING_SHIFT_DAY
+from preprocessing import xml_loader
+from preprocessing.xml_loader import *
+
+
+def already_daily_off_shift(root, employee_offset, employee_rest, day):
+    """
+    Checks if an employee specific day has a interval without demand that exceeds the daily rest
+    hours. If so, there is no need to ensure that daily rest is satisfied, as this is ensured
+    naturally. Returns True or False.
+    """
+
+    no_demand_periods = combine_no_demand_intervals(root)
+    days = get_days(root)
+
+    for interval in no_demand_periods:
+        if day == len(days) - 1 and interval == no_demand_periods[-1] and interval[1] == 24 * len(days):
+            temp_val = interval[1]
+            interval[1] = temp_val + employee_offset
+        if employee_offset + (24 * int(day)) <= interval[0] <= employee_offset + (24 * (int(day) + 1)):
+            if interval[1] - interval[0] >= employee_rest:
+                if interval[0] + employee_rest <= employee_offset + (24 * (int(day) +1)):
+                    return True
+        if interval[0] > employee_offset + (24 * (int(day) + 1)):
+            break
+
+    return False
+
+
+def get_shifts(root):
+
+    demand_intervals = combine_demand_intervals(root)
+    shifts = tuplelist()
+
+    for intervals in demand_intervals:
+
+        if len(intervals) == 2:
+            start_time = intervals[0]
+            duration = intervals[1] - start_time
+
+            if duration >= ALLOWED_SHIFT_DURATION[1]:
+                shifts.append((start_time, duration))
+            else:
+                shifts.append((start_time, intervals[1] - start_time))
+
+        else:
+
+            for time in intervals:
+                found_shift = False
+
+                for t in intervals[intervals.index(time):]:
+                    duration = t - time
+                    if ALLOWED_SHIFT_DURATION[0] <= duration <= ALLOWED_SHIFT_DURATION[1]:
+                        shifts.append((time, duration))
+                        found_shift = True
+                    if duration > ALLOWED_SHIFT_DURATION[1] and not found_shift:
+                        shifts.append((time, duration))
+    return shifts
+
+
+def get_short_and_long_shifts(shifts):
+
+    short_shifts = tuplelist()
+    long_shifts = tuplelist()
+
+    for s in shifts:
+        if s[1] < DESIRED_SHIFT_DURATION[0]:
+            short_shifts.append(s)
+        elif s[1] > DESIRED_SHIFT_DURATION[1]:
+            long_shifts.append(s)
+    return long_shifts, short_shifts
+
+
+def get_shifts_per_day(shifts, time_sets):
+
+    shifts_per_day = tupledict()
+
+    for day in time_sets["days"]:
+        shifts_per_day[day] = []
+
+        for shift in shifts:
+
+            if (
+                    24 * (int(day) - 1) + TIME_DEFINING_SHIFT_DAY
+                    <= shift[0]
+                    < 24 * int(day) + TIME_DEFINING_SHIFT_DAY
+            ):
+                shifts_per_day[day].append(shift)
+
+            if shift[0] >= 24 * int(day) + TIME_DEFINING_SHIFT_DAY:
+                if day == time_sets["days"][-1]:
+                    shifts_per_day[day].append(shift)
+                break
+    return shifts_per_day
+
+
+def get_shifts_violating_daily_rest(root, staff, shifts_per_day):
+    """
+    Returns two values:
+        * violating_shifts: Returns a dict, with "employee" as key and another dict as value. The
+        new dict uses "shift" as key and have a list of shifts that violates daily rest for
+        "employee" if "shift" is worked.
+
+        * invalid_shifts:   Returns a dict with "employee" as key and a list of shifts that is not
+        allowed due to daily rest as value.
+    """
+
+    employees = staff["employees"]
+    daily_rest = staff["employee_daily_rest"]
+    daily_offset = staff["employee_daily_offset"]
+    violating_shifts = tupledict()
+    invalid_shifts = tupledict()
+
+    for e in employees:
+        violating_shifts[e] = tupledict()
+        invalid_shifts[e] = tuplelist()
+        for day in shifts_per_day:
+            if not (already_daily_off_shift(root, daily_offset[e], daily_rest[e], day)):
+                for shift in shifts_per_day[day]:
+                    if day != 0:
+                        for s in shifts_per_day[day - 1]:
+                            if s[0] + s[1] > (24 * int(day)) + daily_offset[e]:
+                                if shift[0] - (s[0] + s[1]) < daily_rest[e] and shift[0] >= (s[0] + s[1]):
+                                    try:
+                                        violating_shifts[e][shift].append(s)
+                                    except:
+                                        violating_shifts[e][shift] = [s]
+                    if day != shifts_per_day.keys()[-1]:
+                        for s in shifts_per_day[day + 1]:
+                            if s[0] < 24 * (int(day) + 1) + daily_offset[e]:
+                                if s[0] - (shift[0] + shift[1]) < daily_rest[e] and s[0] > (shift[0] + shift[1]):
+                                    try:
+                                        violating_shifts[e][shift].append(s)
+                                    except:
+                                        violating_shifts[e][shift] = [s]
+
+                    if shift[0] - (24 * int(day)) - daily_offset[e] < daily_rest[e]:
+                        if 24 * (int(day) + 1) + daily_offset[e] - (shift[0] + shift[1]) < daily_rest[e]:
+                            try:
+                                invalid_shifts[e].append(shift)
+                            except:
+                                invalid_shifts[e] = [shift]
+
+    return [violating_shifts, invalid_shifts]
+
+
+def get_shifts_overlapping_t(shifts, time_sets):
+
+    time_periods = time_sets["periods"][0]
+    shifts_overlapping_t = {}
+
+    for time in time_periods:
+        for shift in shifts:
+            if shift[0] <= time < shift[0] + shift[1]:
+                try:
+                    shifts_overlapping_t[time].append(shift)
+                except:
+                    shifts_overlapping_t[time] = [shift]
+
+    return shifts_overlapping_t
+
+
+def get_off_shifts(root):
+
+    events = get_start_events(root)
+
+    off_shifts = []
+    off_shifts_in_week = tupledict()
+    week = 0
+    off_shifts_in_week[week] = []
+
+    for i in range(len(events)):
+        for event in events[i:]:
+            duration = event - events[i]
+            if events[i] >= (week + 1) * 24 * 7:
+                week += 1
+                off_shifts_in_week[week] = []
+            if event >= (week + 1) * 24 * 7:
+                break
+            if duration > 70:
+                break
+            elif duration >= 36:
+                if (events[i], duration) not in off_shifts:
+                    off_shifts_in_week[week].append((events[i], duration))
+                    off_shifts.append((events[i], duration))
+
+    return [off_shifts, off_shifts_in_week]
+
+
+def get_t_covered_by_off_shifts(off_shifts, time_sets):
+
+    t_covered = tupledict()
+    time_periods = time_sets["periods"][0]
+
+    for shift in off_shifts:
+        end = time_periods.index(shift[0] + shift[1])
+        start = time_periods.index(shift[0])
+        t_covered[shift[0], shift[1]] = time_periods[start:end]
+
+    return t_covered
+
+
+def get_t_covered_by_shift(shifts, time_sets):
+
+    time_step = time_sets["step"]
+    time_periods = time_sets["time_periods"][0]
+    t_covered_by_shift = tupledict()
+
+    for shift in shifts:
+        end = time_periods.index(shift[0] + shift[1] - time_step)
+        start = time_periods.index(shift[0])
+        t_covered_by_shift[shift[0], shift[1]] = time_periods[start:(end + 1)]
+
+    return t_covered_by_shift
+
+
+def get_shift_lookup(shifts_per_day):
+
+    shift_lookup = {}
+
+    for key in shifts_per_day.keys():
+        for value in shifts_per_day[key]:
+            shift_lookup[value] = key
+
+    return shift_lookup
+
+
+def get_shifts_covered_by_off_shifts(shifts, off_shifts):
+
+    shifts_covered = tupledict()
+
+    for off_shift in off_shifts:
+        shifts_covered[off_shift] = []
+        for shift in shifts:
+            if off_shift[0] <= shift[0] < (off_shift[0] + off_shift[1]) or off_shift[0] <= (
+                shift[0] + shift[1]
+            ) < (off_shift[0] + off_shift[1]):
+                shifts_covered[off_shift].append(shift)
+
+    return shifts_covered
+
+
+def load_data(problem_name):
+    root = xml_loader.get_root(problem_name)
+
+    competencies = get_competencies(root)
+    staff = get_employee_lists(root, competencies)
+    time_sets = get_time_sets(root)
+
+    off_shift_sets = get_off_shift_sets(root, time_sets)
+    shift_sets = get_shift_sets(root, staff, time_sets, off_shift_sets["off_shifts"])
+
+    data = {
+        "competencies": competencies,
+        "demand": get_demand(root, competencies),
+        "staff": staff,
+        "limit_on_consecutive_days": 5,
+        "shifts": shift_sets,
+        "off_shifts": off_shift_sets,
+        "time": time_sets,
+        "heuristic": {
+            "t_covered_by_shift": get_t_covered_by_shift(shift_sets["shifts"], time_sets),
+            "shift_lookup": get_shift_lookup(shift_sets["shifts_per_day"]),
+        }
+    }
+
+    return data
+
+
+def get_time_sets(root):
+
+    days = get_days(root)
+    number_of_weeks = int(len(days) / 7)
+
+    return {
+        "step": get_time_steps(root),
+        "periods": get_time_periods(root),
+        "days": days,
+        "weeks": [i for i in range(number_of_weeks)],
+        "saturdays": [5 + i * 7 for i in range(number_of_weeks)],
+        "sundays": [6 + i * 7 for i in range(number_of_weeks)]
+    }
+
+
+def get_off_shift_sets(root, time_sets):
+
+    off_shift_set = get_off_shifts(root)
+
+    return {
+        "t_in_off_shifts": get_t_covered_by_off_shifts(off_shift_set[0], time_sets),
+        "off_shifts": off_shift_set[0],
+        "off_shifts_per_week": off_shift_set[1],
+    }
+
+
+def get_shift_sets(root, staff, time_sets, off_shifts):
+
+    shifts = get_shifts(root)
+
+    shifts_per_day = get_shifts_per_day(shifts, time_sets["days"])
+    long_shifts, short_shifts = get_short_and_long_shifts(shifts)
+    shifts_violating_daily_rest = get_shifts_violating_daily_rest(root, staff, shifts_per_day)
+
+    return {
+        "shifts": shifts,
+        "shifts_per_day": shifts_per_day,
+        "short_shifts": short_shifts,
+        "long_shifts": long_shifts,
+        "shifts_overlapping_t": get_shifts_overlapping_t(shifts, time_sets),
+        "shifts_covered_by_off_shift": get_shifts_covered_by_off_shifts(shifts, off_shifts),
+        "shifts_combinations_violating_daily_rest": shifts_violating_daily_rest[0],
+        "invalid_shifts_violating_daily_rest": shifts_violating_daily_rest[1],
+    }
